Project Management Tool for Freelancer
A freelancer wants to keep track of how much time he spends working on different
projects.
At any given time he is working on several different projects, giving each of them a
unique name. He needs to be able to log the date, the time he starts, and the time he stops
working on a project, so that he knows how many hours in total he has spent on it. He needs
to be able to start and stop the same project multiple times.